% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_multip_seq.R
\name{edge_multip_seq}
\alias{edge_multip_seq}
\title{Edge multiplicity sequences}
\usage{
edge_multip_seq(adj, type = "multigraph")
}
\arguments{
\item{adj}{Integer matrix}

\item{type}{Equals 'graph' if adjacency matrix is for graphs (default),
equals 'multigraph' if it is the equivalence of the adjacency matrix for multigraphs
(with the matrix diagonal double counted)}
}
\value{
All unique edge multiplicity sequences as a dataframe
}
\description{
Finds all unique multigraphs represented  by their edge multiplicity sequences
}
\details{
To be completed
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. *Journal of Social Structure*, 16.
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. *The Journal of Mathematical Sociology*, 40(4), 239-264.
}
\author{
Termeh Shafie
}
