% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degree_seq.R
\name{get_degree_seq}
\alias{get_degree_seq}
\title{Degree sequence of multigraph}
\usage{
get_degree_seq(adj, type = "multigraph")
}
\arguments{
\item{adj}{integer matrix}

\item{type}{equals 'graph' if adjacency matrix is for graphs (default),
equals 'multigraph' if it is the equivalence of adjacency matrix for multigraphs
(with diagonal double counted)}
}
\value{
vector of integers representing the degree sequence
}
\description{
gives the degree sequence of observed graph or multigraph
}
\details{
to be completed
}
\examples{

}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. *Journal of Social Structure*, 16.

Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. *The Journal of Mathematical Sociology*, 40(4), 239-264.
}
\author{
Termeh Shafie
}
