% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_assignment_probs.R
\name{get_edge_assignment_probs}
\alias{get_edge_assignment_probs}
\title{Edge assignment probabilities under IEA model given fixed degrees}
\usage{
get_edge_assignment_probs(m, deg.seq, model)
}
\arguments{
\item{m}{integer giving number of edges in multigraph}

\item{deg.seq}{vector of integers with the sum equal to \emph{2m} representing the degree sequence of the multigraph}

\item{model}{character string, either IEAS or ISA}
}
\value{
A numeric vector representing the edge assignment probabilities
to all possible vertex pair sites
}
\description{
Calculates the edge assignment probabilities given specified degree sequence under the two ways in which the RSM
model can be approximated by the IEA model: the IEAS (independent edge assignment of stubs) or
the ISA (independent stub assignment) model.
}
\details{
The IEAS and ISA edge assignment probabilities to
possible vertex pairs are calculated given a fixed degree sequence \emph{deq.seq}
under the IEAS model, and \emph{deg.seq/2m} under the ISA model.
Number of possible vertex pair sites to which edges are assigned is given by
\eqn{(n+1)n/2} where \emph{n} is number of vertices.
}
\examples{
## Under the IEAS model with 10 possible vertex pair sites (4 vertices)
get_edge_assignment_probs(m = 8, deg.seq = c(4,4,4,4), model = 'IEAS')

## Under the ISA model with 21 possible vertex pair sites (6 vertices)
get_edge_assignment_probs(m = 10, deg.seq = c(8,4,2,2,2,2), model = 'ISA')
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.
}
\author{
Termeh Shafie
}
