% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_multip_seq.R
\name{get_edge_multip_seq}
\alias{get_edge_multip_seq}
\title{Edge multiplicity sequences of multigraphs of a specified size}
\usage{
get_edge_multip_seq(adj, type = "multigraph")
}
\arguments{
\item{adj}{Matrix of integers representing graph adjacency matrix}

\item{type}{Equals 'graph' if adjacency matrix is for graphs (default),
equals 'multigraph' if it is the equivalence of the adjacency matrix for multigraphs
(with matrix diagonal representing loops double counted).}
}
\value{
All unique edge multiplicity sequences as a data frame
}
\description{
Finds all unique multigraphs represented  by their edge multiplicity sequences
}
\details{
Finds all multigraphs as represented by their edge multiplicity sequence
at vertex pair sites. Only practical for small multigraphs.
}
\examples{
## Adjacency matrix for undirected network with 3 nodes
 A <-  matrix(c(0, 1, 2,
                1, 2, 1,
                2, 1, 2), nrow=3, ncol=3)
 get_edge_multip_seq(adj = A, 'multigraph')
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.
}
\author{
Termeh Shafie
}
