% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_multigraph.R
\name{gof_multigraph}
\alias{gof_multigraph}
\title{Goodness of Fit Tests for Random Multigraph Models}
\usage{
gof_multigraph(m, model, deg.mod, hyp, deg.hyp)
}
\arguments{
\item{m}{number of edges}

\item{model}{= assumed model, either RSM, IEAS or ISA}

\item{deg.mod}{= modelled degree sequence}

\item{hyp}{= hypothesis, either IEAS or ISA}

\item{vechyp}{= hypothetical degree sequence
hypothetical degree sequence:
 if 'IEA': d = deg.hyp fully specified
 if 'ISA': p = deg.hyp/2m fully specified
 if 'IEA' and deg.hyp = 0, d estimated from edge mutliplicity sequence
 if 'ISA' and deg.hyp = 0, p estimated from edge mutliplicity sequence}
}
\value{
to be completed
}
\description{
Goodness of fits tests of multigraph models obtained by random stub matching (RSM) and  by independent edge assignments (IEA),
where the latter is either independent edge assignments of stubs (IEAS) or independent stub assignment
}
\details{
to be completed
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. *Journal of Social Structure*, 16.
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. *The Journal of Mathematical Sociology*, 40(4), 239-264.
}
\author{
Termeh Shafie
}
