% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_sim.R
\name{gof_sim}
\alias{gof_sim}
\title{Goodness of fit test simulations}
\usage{
gof_sim(m, model, deg.mod, hyp, deg.hyp)
}
\arguments{
\item{m}{integer giving number of edges in multigraph}

\item{model}{character string representing assumed model, either RSM, IEAS or ISA}

\item{deg.mod}{vector of integers with sum \eqn{2m} representing
the modeled degree sequence of the multigraph}

\item{hyp}{character string representing the hypothesis, either IEAS or ISA}

\item{deg.hyp}{vector of integers with sum equal to \eqn{2m} representing the hypothetical
degree sequence of the multigraph: \cr
  - if 'IEAS': simple IEAS hypothesis with fully specified degree sequence deg.hyp\cr
  - if 'ISA': simple ISA hypothesis with with fully specified stub assignment probabilities deg.hyp/2m\cr
  - if 'IEAS': and deg.hyp = 0: composite IEAS hypothesis with edge multiplicity sequence estimated from data\cr
  - if 'ISA' and deg.hyp = 0: composite ISA hypothesis with edge multiplicity sequence estimated from data\cr}
}
\value{
\item{test.summary}{Expected value and variances of test statistics (stat),
 critical values (cv) according to asymptoticχ²-distribution and
 according to cdf's of test statistics,
 significance level (alpha) according to
 asymptotic χ² distribution, and power of tests (P(stat>cv))}
 \item{degrees.of.freedom}{Degrees of freedom for tests performed}
 \item{probS}{Probability distributions of Pearson statistic S}
 \item{probA}{Probability distributions of information divergence statistic A}
 \item{adjusted.stats}{Expected values and variances for adjusted test statistics, preferred adjusted statistics}
 \item{adjusted.chi2}{Degrees of freedom for adjusted  χ²-distribution}
 \item{power.apx}{Power approximations according to adjusted statistics}
}
\description{
Goodness of fits test simulations of specified multigraph models using Pearson (S) and
information divergence (A) test statistics under the random stub matching (RSM)
and the independent edge assignments (IEA) model,
where the latter is either independent edge assignments of stubs (IEAS) or
independent stub assignment (ISA). These can be used to examine to
check the reliability of the tests by examining the exact probability
distributions of the test statistics and their fit to the asymptotic  χ²-distributions
Only practical for small multigraphs.
}
\details{
The tests are performed using goodness-of-fit measures between simulated
edge multiplicity sequence of a multigraph according to an RSM or IEA model,
and the expected multiplicity sequence according to a simple or composite IEA hypothesis.
Test statistics of Pearson type (S) and of information divergence (A) type are used and summary
of tests given these two statistics are given as output.
}
\examples{
## Testing a simple IEAS hypothesis with degree sequence [6,6,6,2] against
# an IEAS model with degree sequence [14,2,2,2] on a multigraph with n = 4 nodes and m = 10 edges
deg.mod <- c(14,2,2,2)
deg.hyp <- c(6,6,6,2)
test1 <- gof_sim(10, 'IEAS', deg.mod, 'IEAS', deg.hyp)

# Non-null distributions (pdf's and cdf's) of test statistics S and A are given by
test1$probS
test1$probA

# Testing a composite IEAS hypothesis with degree sequence [15,15,15,15] against
# an RSM model with degree sequence [15,15,15,15] on a multigraph with n = 4 nodes and m = 30 edges.
deg.mod <- c(15,15,15,15)
deg.hyp <- c(15,15,15,15)
test2 <- gof_sim(30, 'RSM', deg.mod, 'IEAS', deg.hyp)

# Summary of above tests
test1$test.summary
test2$test.summary

# Non-null distributions (pdf's and cdf's) of test statistics S and A are given by
test2$probS
test2$probA

}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.
}
\seealso{
[gof_stats],[get_edge_assignment_probs],[nsumk],[rsm_model],[gof_test]
}
\author{
Termeh Shafie
}
