% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_stats.R
\name{gof_stats}
\alias{gof_stats}
\title{Moments of Goodness of Fit Statistics}
\usage{
gof_stats(m, dof, m.seq, prob.mg, Q.seq)
}
\value{

}
\description{
Calculates the probability distribution and moments of Peearson's and divergence 
goodness of fit statistics when testing multigraph hypotheses against random multigraoh models.
}
\details{
To be completed
}
\examples{

}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. *Journal of Social Structure*, 16.
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. *The Journal of Mathematical Sociology*, 40(4), 239-264.
}
\author{
Termeh Shafie
}
