% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_test.R
\name{gof_test}
\alias{gof_test}
\title{Goodness of fit tests for multigraph representations of observed networks}
\usage{
gof_test(adj, type, hyp, deg.hyp, dof)
}
\arguments{
\item{adj}{Matrix of integer representing graph adjacency matrix}

\item{type}{Equals 'graph' if adjacency matrix is for graphs (default)}

\item{hyp}{character string representing the hypothesized model (null), either IEAS or ISA}

\item{deg.hyp}{vector of integers with sum equal to 2m representing the null
degree sequence of the multigraph: \cr
  - if 'IEAS': simple IEAS hypothesis with fully specified degree sequence deg.hyp\cr
  - if 'ISA': simple ISA hypothesis with with fully specified stub assignment probabilities deg.hyp/2m\cr
  - if 'IEAS': and deg.hyp = 0: composite IEAS hypothesis with edge multiplicity sequence estimated from data\cr
  - if 'ISA' and deg.hyp = 0: composite ISA hypothesis with edge multiplicity sequence estimated from data\cr}

\item{dof}{integer giving degrees of freedom of test,
r-1 for simple hypotheses and r-n for composite hypotheses where $r = \binom{n+1}{2}$}

\item{m}{integer giving number of edges in multigraph}
}
\value{

}
\description{
Goodness of fits between observed multigraph and specified RSM or IEAS hypotheses
using using Pearson (S) and information divergence (A) tests statistics
}
\details{
The tests are performed using goodness-of-fit measures between the
edge multiplicity sequence of an observed multigraph,
and the expected multiplicity sequence according to a simple or composite IEAS hypothesis.
Test statistics of Pearson (S) and of information divergence (A) type are used and summary
of test using these two statistics are  given as output.
}
\examples{
## adjacency matrix of observed network (multigraph), n = 4 nodes , m = 15 edges
adj <- t(matrix(c( 0, 1, 0, 3,
                   0, 0, 1, 6,
                   0, 1, 0, 3,
                   3, 6, 3, 2), nrow= 4, ncol=4))
# Testing a simple IEAS hypothesis with degree sequence [4 4 8 14]
gof_test(adj, type = 'multigraph', 'IEAS', c(4,4,8,14), 9)
# Testing a composite IEAS hypothesis
gof_test(adj, type  = 'multigraph', 'IEAS', 0, 6)
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. Journal of Social Structure, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. The Journal of Mathematical Sociology, 40(4), 239-264.
}
\seealso{
[get_edgemultipl_seq],[get_degree_seq],[edge_assignment_probs]
}
\author{
Termeh Shafie
}
