% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsumk.R
\name{nsumk}
\alias{nsumk}
\title{Ordered n-tuples of non-negative integers summing to k}
\usage{
nsumk(n, k)
}
\arguments{
\item{n}{A positive integer}

\item{k}{A positive integer}
}
\value{
A matrix X with nchoosek(k+n-1,n-1) rows and n columns. Each row comprises non-negative integers summing to k.
}
\description{
Finds ordered n-tuples of non-integers summing to k.
Only practical for n < 15
}
\details{
To be completed
}
\examples{

}
\author{
Termeh Shafie
}
