% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsm_apx.R
\name{rsm_apx}
\alias{rsm_apx}
\title{IEA approximation of the complexity statistics under the RSM model}
\usage{
rsm_apx(deg.seq)
}
\arguments{
\item{deg.seq}{vector of integers representing the degree sequence of the multigraph}
}
\value{
List including number of multigraphs under the IEA approximation,
summary and interval estomates for number of loops and number of non-loos (M_1 and M__2),
summary and interval estimates for frequencies of edge multiplicites R_k
}
\description{
Summary of statistics estimated for
global structure of random multigraphs under an IEA approximation of the RSM model.
The degree sequence is used to estimate the edge assignment probabilities.
}
\details{
To be completed
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. *Journal of Social Structure*, 16.
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. *The Journal of Mathematical Sociology*, 40(4), 239-264.
}
\author{
Termeh Shafie
}
