% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_assignment_probs.R
\name{edge_assignment_probs}
\alias{edge_assignment_probs}
\title{Edge Assignment Probabilities under the IEAS and ISA model}
\usage{
edge_assignment_probs(m, deg.seq, model)
}
\value{
A vector Q.seq giving the edge assignment probabilities to all possible vertex pair sites
}
\description{
Calculates the edge assignment probabilities given a degree sequence under the two ways in which the RSM
model can be approxiated by the IEA model. This is done by either the IEAS (independent edge assignment of stubs) or
the ISA (independent stub assignment) model.
}
\details{
To be completed
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. *Journal of Social Structure*, 16.
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. *The Journal of Mathematical Sociology*, 40(4), 239-264.
}
\author{
Termeh Shafie
}
