% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_assignment_probs.R
\name{edge_assignment_probs}
\alias{edge_assignment_probs}
\title{Edge Assignment Probabilities under the IEAS and ISA model}
\usage{
edge_assignment_probs(m, deg.seq, model)
}
\arguments{
\item{m}{integer number of edges}

\item{deg.seq}{vector of integers with sum 2m representing the degree sequence of the multigraph}

\item{model}{either}
}
\value{
A vector Q.seq representing the edge assignment probabilities to all possible vertex pair sites
}
\description{
Calculates the edge assignment probabilities given a degree sequence under the two ways in which the RSM
model can be approxiated by the IEA model. This is done by either the IEAS (independent edge assignment of stubs) or
the ISA (independent stub assignment) model.
}
\details{
The IEAS and ISA edge assignment probabilities, to
all possible vertext pairs are calculated given a fixed degree sequence \emph{deq.seq}
under IEAS and \emph{deg.seq/2m} under ISA. Number of possible vertex pair sites is given by
\eqn{(n+1)n/2} where \emph{n} is number of vertices.
}
\examples{
## Under the IEAS model with 10 possible vertex pair sites (4 vertices)
edge_assignment_probs(m = 8, deg.seq = c(4,4,4,4), model = 'IEAS')

#' ## Under the ISA model with 21 possible vertex pair sites (6 vertices)
edge_assignment_probs(m = 10, deg.seq = c(8,4,2,2,2,2), model = 'ISA')
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.
}
\author{
Termeh Shafie
}
