% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degree_seq.R
\name{get_degree_seq}
\alias{get_degree_seq}
\title{Degree sequence of a multigraph}
\usage{
get_degree_seq(adj, type = "multigraph")
}
\arguments{
\item{adj}{Matrix of integers}

\item{type}{Equals 'graph' if adjacency matrix is for graphs (default),
equals 'multigraph' if it is the equivalence of the adjacency matrix for multigraphs
(with the elements of the matrix diagonal double counted)}
}
\value{
Vector of integers representing the degree sequence
}
\description{
Gives the degree sequence of the adjancecy matrix of an observed graph or multigraph
}
\details{
Gives the degree sequence of the adjancecy matrix of an observed graph or multigraph
}
\examples{
 ## Generate random adjancency matrix (6 nodes)
 A <- matrix(sample(c(0,1,2), 36, replace = TRUE), 6, 6, byrow=TRUE)
 # If A represents for graph
 deg.seq <- get_degree_seq(adj = A, type = 'graph' )
 #If A represents a multigraph
 deg.seq <- get_degree_seq(adj = A, type = 'multigraph')
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.
}
\author{
Termeh Shafie
}
