% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_multigraph.R
\name{gof_multigraph}
\alias{gof_multigraph}
\title{Goodness of Fit Tests for Random Multigraph Models}
\usage{
gof_multigraph(m, model, deg.mod, hyp, deg.hyp)
}
\arguments{
\item{adj}{integer matrix}

\item{type}{equals 'graph' if adjacency matrix is for graphs (default),
equals 'multigraph' if it is the equivalence of adjacency matrix for multigraphs
(with diagonal double counted)
model = either RSM, IEAS or ISA with modelled degree sequence deg.mod
vechyp= hypothetical degree sequence,
hypothetical degree sequence:
  if 'IEA': d = deg.hyp fully specified
  if 'ISA': p = deg.hyp/2m fully specified
  if 'IEA' and deg.hyp = 0, d estimated from edge mutliplicity sequence
  if 'ISA' and deg.hyp = 0, p estimated from edge mutliplicity sequence}
}
\value{
vector of integers representing the degree sequence of the multigraph
}
\description{
gives the degree sequence of observed graph or multigraph
}
\details{
to be completed
}
\examples{

}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. *Journal of Social Structure*, 16.
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. *The Journal of Mathematical Sociology*, 40(4), 239-264.
}
\author{
Termeh Shafie
}
