% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_stats.R
\name{gof_stats}
\alias{gof_stats}
\title{Probability distributions of goodness of fit test statistics and summary of tests}
\usage{
gof_stats(m, dof, m.seq, prob.mg, Q.seq)
}
\arguments{
\item{m}{integer giving number of edges in multigraph}

\item{dof}{integer giving degrees of freedom of test}

\item{m.seq}{vector of integers representing all possible multigraphs}

\item{prob.mg}{probability distribution of multigraphs under specified model}

\item{Q.seq}{A numeric vector representing the edge assignment probabilities
to all possible vertex pair sites}
}
\value{
\item{probS}{Probability distribution of Pearson statistic S}
 \item{probA}{Probability distribution of information divergence statistic A}
 \item{summary}{Expected value and variances of test statistics, critical values and significance level
 according to asymptotic chi2-distribution, power of tests}
 \item{adjusted.stats}{Expected value and variances for adjusted test statistics, preferred adjusted statistics}
 \item{adjusted.chi2}{Degrees of freedom for adjusted chi2-distribution}
 \item{power.apx}{Power approximations according to adjusted statistics}
 \item{degrees.of.freedom}{Degrees of freedom for test performed}
}
\description{
Goodness of fits tests of multigraph models using Pearson (S) and information divergence (A) tests statistics
under the random stub matching (RSM) and by independent edge assignments (IEA) model,
where the latter is either independent edge assignments of stubs (IEAS) or independent stub assignment (ISA)
}
\details{
The tests are performed using goodness-of-fit measures between the
edge multiplicity sequence of an observed multigraph,
and the expected multiplicity sequence according to a simple or composite hypothesis.
Test statistics of Pearson type (S) and of information divergence (A) type are used and summary
of tests given these two statistics are given as output.
}
\examples{
## For examples see gof_multigraph()
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. Journal of Social Structure, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. The Journal of Mathematical Sociology, 40(4), 239-264.
}
\seealso{
[gof_multigraph],[edge_assignment_probs],[nsumk],[rsm_model]
}
\author{
Termeh Shafie
}
