% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_stats.R
\name{gof_stats}
\alias{gof_stats}
\title{Exact probability distributions and moments of goodness of fit test statistics}
\usage{
gof_stats(m, dof, m.seq, prob.mg, Q.seq)
}
\arguments{
\item{m}{integer giving number of edges in multigraph}

\item{dof}{integer giving degrees of freedom of test}

\item{m.seq}{vector of integers, each representing possible multigraphs}

\item{prob.mg}{probability distribution of multigraphs under specified model}

\item{Q.seq}{A numeric vector representing the edge assignment probabilities
to all possible vertex pair sites}
}
\value{
\item{test.summary}{Expected value and variances of test statistics (stat),
 critical values (cv) according to asymptotic chi2 distribution and
 according to cdf's of test statistics,
 significance level (alpha) according to
 asymptotic chi2 distribution, and power of tests (P(stat>cv))}
 \item{degrees.of.freedom}{Degrees of freedom for tests performed}
 \item{probS}{Probability distributions of Pearson statistic S}
 \item{probA}{Probability distributions of information divergence statistic A}
 \item{adjusted.stats}{Expected values and variances for adjusted test statistics, preferred adjusted statistics}
 \item{adjusted.chi2}{Degrees of freedom for adjusted chi2-distribution}
 \item{power.apx}{Power approximations according to adjusted statistics and non-central chi2-distributions}
}
\description{
Goodness of fit between observed and expected edge multiplicities, where expected are
calculated under some random  multigraph models using Pearson (S) and information divergence (A) tests statistics.
The exact distribution of these statistics and their moments are calculated using this function.
}
\details{
The tests are performed using goodness-of-fit measures between simulated
edge multiplicity sequence of a multigraph according to an RSM or IEA model,
and the expected multiplicity sequence according to a simple or composite IEA hypothesis.
Test statistics of Pearson type (S) and of information divergence (A) type are used and summary
of tests given these two statistics are given as output.
}
\examples{
## For examples see gof_sim() or gof_test()
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. Journal of Social Structure, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. The Journal of Mathematical Sociology, 40(4), 239-264.
}
\seealso{
[gof_multigraph],[edge_assignment_probs],[nsumk],[rsm_model]
}
\author{
Termeh Shafie
}
