% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea_model.R, R/iea_model2.R
\name{iea_model}
\alias{iea_model}
\title{Complexity statistics under the IEA model for multigraphs}
\usage{
iea_model(
  adj,
  type = "multigraph",
  model = "IEAS",
  K = 0,
  apx = FALSE,
  p.seq = 0
)

iea_model(
  adj,
  type = "multigraph",
  model = "IEAS",
  K = 0,
  apx = FALSE,
  p.seq = 0
)
}
\arguments{
\item{adj}{Matrix of integers representing graph adjacency matrix}

\item{type}{Equals 'graph' if adjacency matrix is for graphs (default),
equals 'multigraph' if it is the equivalence of the adjacency matrix for multigraphs
(with matrix diagonal representing loops double counted)}

\item{model}{character string representing which IEA model: either 'IEAS' (default) or 'ISA'}

\item{K}{Upper limit for \emph{k} in the complexity statistics \emph{R(k)} representing the sequence of
frequencies of vertex pair sites with edge multiplicities \emph{0,1,...,k}. Default is maximum observed in adjacency matrix.v}

\item{apx}{logical (default = 'FALSE'). if 'TRUE', the IEA model is used to approximate
the statistics under the random stub matching model given observed degree sequence}

\item{p.seq}{if model = ISA and apx = FALSE, specify this numerical vector of stub assignment probabilities}
}
\value{
\item{nr.multigraphs}{Number of unique multigraphs possible}
\item{M}{Summary and interval estimates for \emph{number of loops} and \emph{number of multiple edges} (\emph{M1} and \emph{M2})).}
\item{R}{Summary and interval estimates for frequencies of edge multiplicities \emph{R(k)}.}

\item{nr.multigraphs}{Number of unique multigraphs possible}
\item{M}{Summary and interval estimates for \emph{number of loops} and \emph{number of multiple edges} (\emph{M1} and \emph{M2})).}
\item{R}{Summary and interval estimates for frequencies of edge multiplicities \emph{R(k)}.}
}
\description{
Summary of estimated statistics for analyzing
global structure of random multigraphs under the independent edge assignment model
given observed adjacency matrix.

Summary of estimated statistics for analyzing
global structure of random multigraphs under the independent edge assignment model
given observed adjacency matrix. Two versions of the IEA model can be used,
both of which can be used to approximate the RSM model:
\cr
  - independent edge assignment of stubs (IEAS) \cr
  - independent stub assignment (ISA) \cr
}
\details{
The edge assignment probabilities are estimated by using the observed edge multiplicities
(maximum likelihood estimates) or by using the observed degree sequence if the IEA model is used
as an approximation to the RSM model.

When using the IEAS model: the edge assignment probabilities are estimated by using the observed edge multiplicities
(maximum likelihood estimates) or by using the observed degree sequence if the IEA model is used
as an approximation to the RSM model. \cr
When using the ISA model: The stub assignment probabilities are estimated by using
the observed degree sequence over \emph{2m} if the ISA model is used
as an approximation to the RSM model. Otherwise, a sequence containing the stub assignment
probabilities (for example based on prior belief) should be given as argument.
}
\examples{
## Adjacency matrix of a small graph on 3 nodes
A <-  matrix(c(1, 1, 0,
               1, 2, 2,
               0, 2, 0),
             nrow = 3, ncol = 3)
iea_model(adj = A , type = 'graph', K = 0, apx = FALSE)
## Adjacency matrix of a small graph on 3 nodes
A <-  matrix(c(1, 1, 0,
               1, 2, 2,
               0, 2, 0),
             nrow = 3, ncol = 3)
## When the IEAS model is used
iea_model(adj = A , type = 'graph', model = 'IEAS', K = 0, apx = FALSE)

## When the IEAS model is used to approximate the RSM model
iea_model(adj = A , type = 'graph', model = 'IEAS', K = 0, apx = TRUE)

## When the ISA model is used to approximate the RSM model
iea_model(adj = A , type = 'graph', model = 'ISA', K = 0, apx = TRUE)

## When the ISA model is used with a pre-specified stub assignment probabilities
iea_model(adj = A , type = 'graph', model = 'ISA', K = 0, apx = FALSE, p.seq = c(1/3, 1/3, 1/3))
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.

Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.
}
\seealso{
[get_degree_seq]

[get_degree_seq]
}
\author{
Termeh Shafie

Termeh Shafie
}
