% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea_model.R
\name{iea_model}
\alias{iea_model}
\title{Complexity statistics under the IEA model for multigraphs}
\usage{
iea_model(adj, type = "multigraph", K = 0, apx = FALSE)
}
\arguments{
\item{adj}{Matrix of integers}

\item{type}{Equals 'graph' if adjacency matrix is for graphs (default),
equals 'multigraph' if it is the equivalence of the adjacency matrix for multigraphs
(with the matrix diagonal double counted)}

\item{K}{Upper limit for k in the complexity statistics R_k. Default is maximum observed in adjacency matrix}

\item{apx}{Logical (default = 'FALSE'). if 'TRUE', the IEA model is used to approximate
the statistics under the random stub matching model given observed degree sequence}
}
\value{

}
\description{
Summary of statistics estimated for
global structure of random multigraphs under independent edge assignment model
given observed adjacency matrix.
The edge assignment probabilities are estimated using the observed edge multiplicities
(maximum likelihood estimates)
}
\details{
To be completed
}
\examples{

}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.
}
\seealso{
[get_degree_seq]
}
\author{
Termeh Shafie
}
