% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsm_model.R
\name{rsm_model}
\alias{rsm_model}
\title{Random stub matching model for multigraphs}
\usage{
rsm_model(deg.seq)
}
\arguments{
\item{deg.seq}{Vector of integers representing the degree sequence of a multigraph}
}
\value{
\item{m.seq}{Possible multigraphs represented by edge multiplicity sequences}
\item{prob.dists}{Probability distrbution of the multigraphs/edge multiplicity sequences,
statistics 'number of loops', 'number of multiple edges', and 'simple graph' (logical) for each multigraph}
\item{M}{Summary and interval estimates for 'number of loops' and 'number of multiple edges' (\emph{M1} and \emph{M2})).}
}
\description{
Given the degree sequence of a multigraph,
this function finds all unique multigraphs represented
by their edge multiplicity sequences and different
complexity statistics together with their probability distributions and moments.
}
\details{
To be completed
}
\examples{
D <- c(2,2,3,3) # degree sequence
mod1 <- rsm_model(c(2,2,3,3))
mod1$m.seq
mod1$prob.dists
mod1$stat.moms
}
\references{
Shafie, T. (2015). A Multigraph Approach to Social Network Analysis. \emph{Journal of Social Structure}, 16.
\cr
Shafie, T. (2016). Analyzing Local and Global Properties of Multigraphs. \emph{The Journal of Mathematical Sociology}, 40(4), 239-264.
}
\author{
Termeh Shafie
}
